Explode into multiple classes, replac print with toString
